updates for assignment 14
