+++ week14/assignment/runs.txt
@@ -1,0 +1,3 @@
+1,6
+2,6
+3,6
