+++ week14/prove.md
@@ -8,7 +8,7 @@ You will be implementing both a [depth-first algorithm](https://www.youtube.com/
 
 The assignment is divided into two parts.  The first part will retrieve the family tree information using DFS.  The second part will use BFS.
 
-There are a number of classes in `assignment.py` that you can use:
+There are a number of classes in `common.py` that you can use:
 
 **Person**
 
@@ -24,9 +24,16 @@ The tree class allows you to build a family tree from the families and individua
 
 **Request_thread**
 
-The only method of requesting information from the server.  Do not change it.
+The only method of requesting information from the server.
 
 ## Assignment
+
+### Assignment files
+
+- `functions.py` - This is the Python file that you will be changing and submitting for the assignment.  **THIS IS THE ONLY FILE THAT YOU CAN MODIFY**.
+- `assignment.py` - This file contains main() and will call your functions in the `functions.py` Python file.
+- `common.py` - Contains common classes and functions for the assignment.
+- `runs.txt` - Contains which search algorithms will be run and how many generations will be retrieved from the server.  You can modify this file while you are testing your program.
 
 ### FS Server
 
@@ -35,50 +42,44 @@ There is a server program that you will need to run in it's own terminal window 
 1. Open a command or terminal window on your computer.
 1. Go to the directory of the assignment.
 1. Type `python server.py` to start the server.
+1. Note if the server is very busy it might not reply with a 200 reply code or might return an empty JSON.  You should handle these issues in your program.
 
 While running, the server will display requests and replies from your assignment.  It also displays the current number of active threads making requests and the maximum number of threads.
 
-The project file is `assignment.py` in the `week14/assignment` folder in GitHub.
-
-The file `assignment.py` contains two functions: `part1()` and `part2()`.
-
 ### Part 1
 
-- The function `part1()` currently retrieves the pedigree information one family and person at a time.  Each call to the server takes 0.25 seconds.
+- The function `depth_fs_pedigree()` will retrieve the family tree using a recursive algorithm.
 - Your task is to use threads to make this function faster.  There is no limit on the number of threads you can use.
-- You must build the pedigree tree starting with the starting family id.  You must write your program to handle different family information from the server (ie., number of families, number of children, etc.).
-- The given code in `part1()` uses Depth First Search to retrieve information. You can change it to use threads to go up the husband's line and wife's line concurrently.
-- You must retrieve all individuals in a family (ie., husband, wife and children)
-
-**Your goal is to execute part 1 in under 10 seconds**
+- You must build the pedigree tree starting with the starting family id.  You must write your program to handle different family information from the server (ie., number of families, number of children, etc.).  A family might be missing a parent and the number of children is random.
+- You must retrieve all individuals in a family and add them to the tree object. (ie., husband, wife and children)
+- Suggestion: get the function to work without using threads first.
+- **Your goal is to execute part 1 in under 10 seconds for 6 generations**
 
 ### Part 2
 
-- In this part, you will be retrieving the family information using a breadth-first algorithm.  Refer to [this link](https://www.youtube.com/watch?v=86g8jAQug04) for more information.
-- Use threads to speed up requesting families and individuals from the server.
+- In this part, you will be retrieving the family information using a breadth-first algorithm in the function `breadth_fs_pedigree()`.
 - Your goal is to make this function as fast as possible by using threads.
 - You must retrieve all individuals in a family (ie., husband, wife and children)
 - Do not use recursion for this algorithm.
-
-**Your goal is to execute part 2 in under 10 seconds**
+- **Your goal is to execute part 2 in under 10 seconds for 6 generations**
 
 ### 10% Bonus
 
-- A 10% bonus if you can limit the number of threads to 5 while retrieving the family information in part 2.  The server must not have more than 5 active threads at a time. (watch the output from the server to see number of activity threads)
+- A 10% bonus if you can limit the number of threads to 5 while retrieving the family information in part 2 using function `breadth_fs_pedigree_limit5()`.  
+- The server must not have more than 5 active threads at a time. (watch the output from the server to see number of activity threads)
 - There is no time requirement for this extra feature.
-- Implement the function bonus().
+
 
 ### Misc
 
 - The server will create a log file called `server.log`
 - You assignment program will create a log file named `assignment.log`
-- Do not change these log files.
 
 ## Rubric
 
 **Describing your code**
 
-In the header of the file ``assignment.py` is a place where you can give meaningful descriptions of part 1 and part 2.  If this is missing or incomplete, 10 points will be taken off your grade for this assignment.
+In the header of the file `functions.py` is a place where you can give meaningful descriptions of part 1 and part 2.  If this is missing or incomplete, 10 points will be taken off your grade for this assignment.
 
 Assignments are not accepted late. Instead, you should submit what you have completed by the due date for partial credit.
 
@@ -86,7 +87,7 @@ Assignments are individual and not team based.  Any assignments found to be  pla
 
 ## Submission
 
-- Add comments in the Python file in the header section describing how you sped up the program.
+- Add comments in the `functions.py` Python file in the header section describing how you sped up the program.
 - You must clearly show in your code that you are using DFS and BFS alogrithms.
-- Upload your Python file to I-Learn.
+- Upload only that Python file to I-Learn.
